requirements.txt
absl-py==1.2.0
accelerate==0.3.0
aiohttp==3.8.3
aiosignal==1.2.0
alabaster==0.7.12
antlr4-python3-runtime==4.8
anyascii==0.3.1
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.12.10
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrdict==2.0.1
attrs==20.2.0
audioread==2.1.9
babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.10.0
black==19.10b0
bleach==3.3.1
boto3==1.24.89
botocore==1.27.89
braceexpand==0.1.7
bs4==0.0.1
cachetools==4.2.4
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
click-default-group==1.2.2
coloredlogs==15.0.1
commonmark==0.9.1
configparser==5.0.2
contractions==0.1.72
coverage==6.5.0
cycler==0.10.0
cython==0.29.24
dataclasses==0.6
datasets==2.2.1
debugpy==1.4.0
decorator==5.1.1
defusedxml==0.7.1
detoxify==0.5.0
dill==0.3.5.1
distance==0.1.3
distlib==0.3.6
docformatter==1.4
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
editdistance==0.5.3
emoji==2.1.0
entrypoints==0.3
executing==1.1.1
fairscale==0.4.8
fairseq==0.10.2
filelock==3.8.0
fire==0.4.0
flake8==5.0.4
flake8-bugbear==22.9.23
flask==1.1.4
frozendict==2.0.3
frozenlist==1.3.1
fsspec==2022.2.0
future==0.18.2
fuzzywuzzy==0.18.0
g2p-en==2.1.0
gdown==3.13.0
gevent==22.8.0
gevent-websocket==0.10.1
gitdb==4.0.9
gitdb2==4.0.2
gitpython==3.1.28
google==3.0.0
google-api-core==2.10.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.49.1
h5py==3.3.0
html2text==2020.1.16
html5lib==1.1
huggingface-hub==0.10.0
humanfriendly==10.0
hydra-core==1.1.1
idna==3.4
imagesize==1.4.1
importlib-metadata==4.2.0
importlib-resources==5.2.3
inflect==5.3.0
iniconfig==1.1.1
iopath==0.1.10
ipadic==1.0.0
ipykernel==6.0.3
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.10.1
itsdangerous==1.1.0
jedi==0.18.1
jieba==0.42.1
jinja2==2.11.3
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaldi-io==0.9.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.3.1
latexcodec==2.0.1
lazy-object-proxy==1.7.1
librosa==0.8.1
llvmlite==0.36.0
lxml==4.6.3
markdown==3.3.2
markdown-it-py==0.5.8
markupsafe==2.0.1
marshmallow==3.13.0
matplotlib==3.4.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdutils==1.4.0
mecab-python3==1.0.3
megatron-lm==2.2.0
mephisto==1.0.3
mistune==0.8.4
mock==4.0.3
more-itertools==8.14.0
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.13
mypy==0.971
mypy-extensions==0.4.3
myst-parser==0.12.10
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nemo-toolkit==1.1.0
nest-asyncio==1.5.1
ninja==1.10.2.4
nltk==3.7
notebook==6.4.0
numba==0.53.1
numpy==1.21.0
oauthlib==3.2.1
omegaconf==2.1.1
onnx==1.9.0
opencc==1.1.2
packaging==21.3
pandas==1.5.0
pandocfilters==1.4.3
pangu==4.0.6.1
parameterized==0.8.1
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pdoc==10.0.4
pesq==0.0.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.2.0
pip-api==0.0.20
pipenv==2022.9.24
pipreqs==0.4.10
platformdirs==2.5.2
pluggy==0.13.1
pooch==1.4.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-gfm==1.0.2
py-rouge==1.1
pyahocorasick==1.4.4
pyamg==4.1.0
pyaml==20.4.0
pyannote-core==4.1
pyannote-database==4.1
pyannote-metrics==3.0.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.0
pybtex==0.24.0
pybtex-docutils==1.0.0
pycodestyle==2.9.1
pydeprecate==0.3.0
pydub==0.25.1
pyflakes==2.5.0
pygments==2.13.0
pylint==2.15.3
pyparsing==3.0.9
pypinyin==0.42.0
pyrsistent==0.18.0
pysocks==1.7.1
pystoi==0.3.3
pytest==5.4.3
pytest-cov==2.12.1
pytest-datadir==1.3.1
pytest-regressions==2.3.1
pytest-runner==5.3.1
python-dateutil==2.8.2
pytorch-lightning==1.3.7.post0
pytz==2022.4
pyyaml==5.4.1
pyzmq==24.0.1
rapidfuzz==1.4.1
redis==3.5.3
regex==2022.9.13
requests==2.28.1
requests-mock==1.10.0
requests-oauthlib==1.3.1
resampy==0.2.2
responses==0.18.0
rich==12.6.0
rich-click==1.5.2
rouge-score==0.0.4
rsa==4.9
ruamel-yaml==0.17.10
ruamel-yaml-clib==0.2.6
s3transfer==0.6.0
sacrebleu==1.5.1
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.6.1
send2trash==1.7.1
sentence-transformers==2.1.0
sentencepiece==0.1.97
sentry-sdk==1.3.0
sh==1.14.3
shellingham==1.4.0
shortuuid==1.0.1
simplejson==3.17.3
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.10.3.post1
soupsieve==2.2.1
sox==1.4.1
sphinx==2.2.2
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.3.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.5.1
subprocess32==3.5.4
subword-nmt==0.3.8
sympy==1.8
tabulate==0.8.10
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardx==2.5
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
textsearch==0.0.24
threadpoolctl==3.1.0
timeout-decorator==0.5.0
tokenizers==0.13.1
toml==0.10.2
tomli==1.2.3
tomlkit==0.11.5
torch-stft==0.1.4
torchaudio==0.9.1
torchmetrics==0.3.2
torchtext==0.10.1
tornado==6.2
tqdm==4.62.3
traitlets==5.4.0
transformers==4.24.0
typed-ast==1.4.3
typer==0.3.2
typing-extensions==4.4.0
unidecode==1.3.6
untokenize==0.1.1
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
wandb==0.11.0
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
websocket-client==1.4.1
websocket-server==0.5.6
websockets==10.3
werkzeug==1.0.1
wget==3.2
wheel==0.37.1
widgetsnbextension==3.5.1
wrapt==1.14.1
xmltodict==0.12.0
xxhash==3.0.0
yarg==0.1.9
yarl==1.8.1
youtokentome==1.0.6
zipp==3.9.0
zope-event==4.5.0
zope-interface==5.4.0